requirements.txt
Django==4.2.23
oemof.solph==0.5.6
pandas>=2.2.0
numpy>=1.24.3
openpyxl==3.0.10
gunicorn==21.2.0
whitenoise>=6.5.0
dj-database-url==2.1.0
